Controllers

+++ src/main/java/com/cumulusclouds/w4153cumuluscloudsmsusermanagement/service/MusicianService.java
@@ -5,43 +5,58 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.UUID;
 
 import com.cumulusclouds.w4153cumuluscloudsmsusermanagement.model.Musician;
+import com.cumulusclouds.w4153cumuluscloudsmsusermanagement.model.Account;
 import com.cumulusclouds.w4153cumuluscloudsmsusermanagement.repository.MusicianRepository;
+import com.cumulusclouds.w4153cumuluscloudsmsusermanagement.repository.AccountRepository;
 
 @Service
 @Transactional
 public class MusicianService {
 
-  @Autowired
-  private MusicianRepository musicianRepository;
+    @Autowired
+    private MusicianRepository musicianRepository;
 
-  public List<Musician> getAllMusicians() {
-    return musicianRepository.findAll();
-  }
+    @Autowired
+    private AccountRepository accountRepository;
 
-  public Musician getMusicianById(UUID id) {
-    return musicianRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Musician not found"));
-  }
-
-  public Musician createMusician(Musician musician) {
-    return musicianRepository.save(musician);
-  }
-
-  public Musician updateMusician(UUID id, Musician updatedMusician) {
-    Musician existingMusician = getMusicianById(id);
-    existingMusician.setGenre(updatedMusician.getGenre());
-    existingMusician.setInstrumentsPlayed(updatedMusician.getInstrumentsPlayed());
-    existingMusician.setYearsOfExperience(updatedMusician.getYearsOfExperience());
-    existingMusician.setSampleWorks(updatedMusician.getSampleWorks());
-    existingMusician.setAvailability(updatedMusician.getAvailability());
-    return musicianRepository.save(existingMusician);
+    public List<Musician> getAllMusicians() {
+        return musicianRepository.findAll();
     }
 
-  public void deleteMusician(UUID id) {
-    musicianRepository.deleteById(id);
-  }
-  
+    public Optional<Musician> getMusicianById(UUID id) {
+        return musicianRepository.findById(id);
+    }
+
+    public Musician createMusician(Musician musician, UUID accountId) {
+        Optional<Account> account = accountRepository.findById(accountId);
+        if (account.isEmpty()) {
+            throw new IllegalArgumentException("Account not found");
+        }
+        musician.setAccount(account.get());
+        return musicianRepository.save(musician);
+    }
+
+    public Optional<Musician> updateMusician(UUID id, Musician musicianDetails) {
+        return musicianRepository.findById(id)
+            .map(musician -> {
+                musician.setGenre(musicianDetails.getGenre());
+                musician.setInstrumentsPlayed(musicianDetails.getInstrumentsPlayed());
+                musician.setYearsOfExperience(musicianDetails.getYearsOfExperience());
+                musician.setSampleWorks(musicianDetails.getSampleWorks());
+                musician.setAvailability(musicianDetails.getAvailability());
+                return musicianRepository.save(musician);
+            });
+    }
+
+    public boolean deleteMusician(UUID id) {
+        if (musicianRepository.existsById(id)) {
+            musicianRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
 }
